+++ apps/python-docs/pages/api/steamship/[...steamship].ts
@@ -1,7 +1,0 @@
-import { createNextPagesApiHandler } from '@steamship/react/next/server';
-
-export const config = {
-  runtime: 'edge'
-};
-
-export default createNextPagesApiHandler();