+++ apps/workshop/app/layout.tsx
@@ -2,6 +2,8 @@ import "./globals.css";
 import "@steamship/react/styles.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
+import { ThemeProvider } from "./components/theme-provider";
+import { ModeToggle } from "./components/mode-toggle";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -16,8 +18,15 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en">
-      <body className={inter.className}>{children}</body>
+    <html lang="en" suppressHydrationWarning>
+      <body className={inter.className}>
+        <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+          <>
+            <ModeToggle />
+            {children}
+          </>
+        </ThemeProvider>
+      </body>
     </html>
   );
 }