+++ examples/with-nextjs-auth-and-database/app/add-pet-button.tsx
@@ -1,0 +1,94 @@
+"use client";
+import { Button, buttonVariants } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "@/components/ui/dialog";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Textarea } from "@/components/ui/textarea";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+
+const AddPetButton = () => {
+  const [open, setOpen] = useState(false);
+  const [name, setName] = useState("");
+  const [breed, setBreed] = useState("");
+  const [description, setDescription] = useState("");
+
+  const router = useRouter();
+
+  const onSave = async () => {
+    await fetch(
+      `/api/add-pet?dogName=${name}&dogDescription=${description}&dogBreed=${breed}`,
+      { method: "POST" }
+    );
+    setBreed("");
+    setName("");
+    setDescription("");
+    setOpen(false);
+    router.refresh();
+  };
+
+  return (
+    <Dialog open={open} onOpenChange={setOpen}>
+      <DialogTrigger asChild>
+        <Button className={buttonVariants()}>Add Dog</Button>
+      </DialogTrigger>
+      <DialogContent className="sm:max-w-[425px]">
+        <DialogHeader>
+          <DialogTitle>Add Dog</DialogTitle>
+          <DialogDescription>
+            Add a new dog to your profile here. Click save when you&apos;re
+            done.
+          </DialogDescription>
+        </DialogHeader>
+        <div className="grid gap-4 py-4">
+          <div className="grid grid-cols-4 items-center gap-4">
+            <Label htmlFor="name" className="text-right">
+              Name
+            </Label>
+            <Input
+              id="name"
+              value={name}
+              onChange={(e) => setName(e.target.value)}
+              className="col-span-3"
+            />
+          </div>
+          <div className="grid grid-cols-4 items-center gap-4">
+            <Label htmlFor="breed" className="text-right">
+              Breed
+            </Label>
+            <Input
+              id="breed"
+              value={breed}
+              onChange={(e) => setBreed(e.target.value)}
+              className="col-span-3"
+            />
+          </div>
+          <div className="grid grid-cols-4 items-center gap-4">
+            <Label htmlFor="description" className="text-right">
+              Description
+            </Label>
+            <Textarea
+              id="description"
+              value={description}
+              onChange={(e) => setDescription(e.target.value)}
+              className="col-span-3"
+            />
+          </div>
+        </div>
+        <DialogFooter>
+          <Button onClick={onSave}>Save</Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+};
+
+export default AddPetButton;