+++ apps/workshop/CHANGELOG.md
@@ -1,5 +1,84 @@
 # workshop
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.13.0
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.12.0
+
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.11.0
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.10.0
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.9.0
+
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @steamship/react@0.8.0
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.7.2
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.7.1
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.7.0
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @steamship/react@0.6.0
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @steamship/react@0.5.0
+
 ## 0.1.6
 
 ### Patch Changes